+++ pkg/config/types.go
@@ -76,7 +76,7 @@ type Config struct {
 	ClusterUid                          string
 	DnsSyncInterval                     time.Duration
 	CrdPath                             string
-	EnableGateway                       bool
+	EnableGatewayTLS                    bool
 	DisableExpensiveCache               bool
 	EnableInternalLogging               bool
 	EnabledWorkloadIdentity             bool
@@ -85,4 +85,5 @@ type Config struct {
 	DefaultDomainServerAddress string
 	DefaultDomainClientID      string
 	DefaultDomainZoneID        string
+	EnableDefaultDomainGateway bool
 }