+++ pkg/controller/controller.go
@@ -17,7 +17,6 @@ import (
 	"github.com/Azure/aks-app-routing-operator/pkg/controller/service"
 	"github.com/Azure/aks-app-routing-operator/pkg/store"
 	"github.com/Azure/aks-app-routing-operator/pkg/util"
-	"github.com/Azure/go-autorest/autorest/azure"
 	"github.com/go-logr/logr"
 	cfgv1alpha2 "github.com/openservicemesh/osm/pkg/apis/config/v1alpha2"
 	policyv1alpha1 "github.com/openservicemesh/osm/pkg/apis/policy/v1alpha1"
@@ -177,7 +176,7 @@ func setupIndexers(mgr ctrl.Manager, lgr logr.Logger, conf *config.Config) error
 		return fmt.Errorf("adding Nginx Ingress Controller IngressClass indexer: %w", err)
 	}
 
-	if conf.EnableGateway {
+	if conf.EnableGatewayTLS {
 		if err := util.AddGatewayServiceAccountIndex(mgr.GetFieldIndexer(), gatewayListenerIndexName); err != nil {
 			lgr.Error(err, "adding Gateway Service Account indexer")
 			return fmt.Errorf("adding Gateway Service Account indexer: %w", err)
@@ -255,7 +254,7 @@ func setupControllers(mgr ctrl.Manager, conf *config.Config, lgr logr.Logger, cl
 		return fmt.Errorf("setting up ingress backend reconciler: %w", err)
 	}
 
-	if conf.EnableGateway {
+	if conf.EnableGatewayTLS {
 		lgr.Info("setting up gateway reconcilers")
 		if err := spc.NewGatewaySecretClassProviderReconciler(mgr, conf, gatewayListenerIndexName); err != nil {
 			return fmt.Errorf("setting up Gateway SPC reconciler: %w", err)
@@ -269,12 +268,6 @@ func setupControllers(mgr ctrl.Manager, conf *config.Config, lgr logr.Logger, cl
 	if conf.EnableDefaultDomain {
 		lgr.Info("setting up default domain reconcilers")
 
-		// Parse the DefaultDomainZoneID to extract subscription, resource group, and cluster name
-		parsedZone, err := azure.ParseResourceID(conf.DefaultDomainZoneID)
-		if err != nil {
-			return fmt.Errorf("parsing default domain zone ID: %w", err)
-		}
-
 		// Create default domain client for fetching TLS certificates
 		defaultDomainClient := defaultdomain.NewCachedClient(
 			context.Background(),
@@ -282,13 +275,6 @@ func setupControllers(mgr ctrl.Manager, conf *config.Config, lgr logr.Logger, cl
 				Opts: defaultdomain.Opts{
 					ServerAddress: conf.DefaultDomainServerAddress,
 				},
-				SubscriptionID: parsedZone.SubscriptionID,
-				ResourceGroup:  parsedZone.ResourceGroup,
-				// note that the zone name isn't actually the cluster name, but the default domain svc resource
-				// doesn't actually utilize this value for anything other than logging/tracing purposes so the unique
-				// zone name works as well. Default-domain-svc only cares about the ccp id
-				ClusterName: parsedZone.ResourceName,
-				CCPID:       conf.ClusterUid,
 			},
 			mgr.GetLogger().WithName("default-domain-client"),
 		)