+++ pkg/manifests/external_dns.go
@@ -32,6 +32,9 @@ const (
 	k8sNameKey              = "app.kubernetes.io/name"
 	externalDnsResourceName = "external-dns"
 	txtWildcardReplacement  = "approutingwildcard"
+
+	// ExternalDNSVersion is the version of the external-dns image used
+	ExternalDNSVersion = "v0.20.0"
 )
 
 type IdentityType int
@@ -56,7 +59,8 @@ const (
 	ResourceTypeIngress ResourceType = iota
 	ResourceTypeGateway
 
-	maxUIDLength = 8
+	maxUIDLength   = 16
+	checkSumLength = 16
 )
 
 func (rt ResourceType) String() string {
@@ -599,7 +603,7 @@ func newExternalDNSDeployment(conf *config.Config, externalDnsConfig *ExternalDn
 
 	podLabels := GetTopLevelLabels()
 	podLabels["app"] = externalDnsConfig.resourceName
-	podLabels["checksum/configmap"] = configMapHash[:maxUIDLength]
+	podLabels["checksum/configmap"] = configMapHash[:checkSumLength]
 
 	if externalDnsConfig.identityType == IdentityTypeWorkloadIdentity {
 		podLabels["azure.workload.identity/use"] = "true"
@@ -658,7 +662,7 @@ func newExternalDNSDeployment(conf *config.Config, externalDnsConfig *ExternalDn
 					ServiceAccountName: serviceAccount,
 					Containers: []corev1.Container{*withLivenessProbeMatchingReadiness(withTypicalReadinessProbe(7979, &corev1.Container{
 						Name:  "controller",
-						Image: path.Join(conf.Registry, "/oss/v2/kubernetes/external-dns:v0.20.0"),
+						Image: path.Join(conf.Registry, "/oss/v2/kubernetes/external-dns:"+ExternalDNSVersion),
 						Args:  deploymentArgs,
 						VolumeMounts: []corev1.VolumeMount{{
 							Name:      "azure-config",